fix(tests): resolve compilation errors in offline functionality tests
- Replace ApplicationProvider.getApplicationContext() with createOnlineContext() helper in ViewModel tests to avoid IllegalStateException
- Remove CompositionLocalProvider with mocked context from HomeScreenStoriesTest to fix ClassCastException with Coil
- Remove manual network setup in Robolectric tests (use default network configuration)
- Fix extra closing brace in HomeScreenStoriesTest after removing withOnlineContext wrapper
- Remove unused imports (ShadowNetworkCapabilities, mockk imports)
